Add documentation for enabling Eloquence on secure and log-on screensrt
This pull request updates the README to include clear instructions on how to enable Eloquence on secure, start-up, and log-on screens when using NVDA 64-bit.
It also addresses and closes issue #31 by providing the missing steps required to manually copy the `eloquence_host32.exe` file to the correct system configuration directory.